Extract text ingress dispatch path
